pypeline: factor off uvicorn utility functions
Factor out uvicorn utility functions so that they can be re-used in all
the places uvicorn is used.
